Basic updater code finished
Still need to do a really test on minecraft itself but in eclipse it
download the update list from dropbox. Then check if all files where
precent. With the list empty it came back as all updated. Which doesn't
sound like a sure thing but reguardless of empty list it still checks
every file in mods folder.