Fixed Downloader saving in the correct location
Well it wasn't just the downloader, but also the ListProcessor was
saving only to the mods folder and not too the backup mods folder. It is
now corrected and any mod using a plain Url download can be downloaded
with this. Mod list also functions as well. Now to just finish off the
GUI and this thing is read to go.